requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
django-rest-auth==0.9.5
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
